backend/requirements.txt
django
python-dotenv
django-cors-headers
djangorestframework
jwt
dj-rest-auth
django-allauth
psycopg2
dj-database-url
Pillow
requests